+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 
-# 1.1
+## 1.1
+
+### Fixes
+- Fixed navigation to 2GIS from the app.
+- Resolved an issue where generating a QR code took too long when the list was large.
+- Fixed vertical scrolling in nested lists.
+- Fixed some design bugs.
+
+### Features
+- Added search inside lists when the number of items exceeds 10.
+- Added Organic Maps as a navigation option.
 
 ## 1.0
-
-